pyproject.toml
[tool.poetry]
name = "ben-xu-bot"
version = "0.1.0"
description = ""
authors = ["listar2000 <35262801+listar2000@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "ben_xu_bot"}]

[tool.poetry.dependencies]
python = "^3.9"
discord-py = "^2.1.1"
python-dotenv = "^0.21.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

